+++ src/components/Post/screens/PostScreen.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useMemo, useRef, useState } from 'react';
+import React, { useCallback, useEffect, useRef, useState } from 'react';
 import { NativeStackNavigationProp } from '@react-navigation/native-stack';
 import {
   setNewCommentId,
@@ -23,6 +23,22 @@ import LoadingScreen from '@components/Common/Loading/LoadingScreen';
 import { YStack } from 'tamagui';
 import RefreshControl from '@components/Common/Gui/RefreshControl';
 
+interface IPostScreenContext {
+  refresh?: () => void;
+  postCollapsed: boolean;
+  setPostCollapsed?: React.Dispatch<React.SetStateAction<boolean>>;
+}
+
+const PostScreenContext: React.Context<IPostScreenContext> =
+  React.createContext<IPostScreenContext>({
+    refresh: undefined,
+    postCollapsed: false,
+    setPostCollapsed: undefined,
+  });
+
+export const usePostScreenContext = (): IPostScreenContext =>
+  React.useContext(PostScreenContext);
+
 interface IProps {
   navigation: NativeStackNavigationProp<any>;
   route: any;
@@ -40,10 +56,13 @@ export default function PostScreen({
   navigation,
   route,
 }: IProps): React.JSX.Element {
-  const { postId, scrollToCommentId } = route.params;
+  const { postId, parentCommentId } = route.params;
 
   const awaitTransition = useAwaitTransition();
 
+  const [postCollapsed, setPostCollapsed] = useState<boolean>(
+    parentCommentId != null,
+  );
   const postCounts = usePostCounts(postId);
   const postCommentsInfo = usePostCommentsInfo(postId);
   const postCommunityId = usePostCommunityId(postId);
@@ -60,15 +79,21 @@ export default function PostScreen({
 
   const flashListRef = useRef<FlashList<ICommentInfo>>();
 
-  const commentsToShow = useMemo(() => {
-    return !awaitTransition.transitioning
-      ? postCommentsInfo?.filter((commentInfo) => commentInfo.showInPost)
-      : [];
-  }, [awaitTransition.transitioning, postCommentsInfo]);
-
   const { isLoading, isError, isRefreshing, refresh } = useLoadData(
     async () => {
-      await instance.getComments(postId, postCommunityId ?? 0, sortType);
+      await instance.getComments({
+        post_id: postId,
+        community_id: postCommunityId ?? undefined,
+        sort: sortType,
+        parent_id: parentCommentId ?? undefined,
+        ...(parentCommentId != null && {
+          max_depth: 10,
+        }),
+      });
+
+      if (parentCommentId != null) {
+        flashListRef.current?.scrollToEnd({ animated: true });
+      }
 
       initialized.current = true;
     },
@@ -95,15 +120,21 @@ export default function PostScreen({
 
     if (!initialized.current) return;
 
+    flashListRef.current?.scrollToOffset({ offset: 0, animated: true });
+
     refresh(async () => {
-      await instance.getComments(postId, postCommunityId ?? 0, sortType);
+      await instance.getComments({
+        post_id: postId,
+        community_id: postCommunityId ?? undefined,
+        sort: sortType,
+      });
     });
   }, [sortType]);
 
   useEffect(() => {
     if (newCommentId == null) return;
 
-    const index = commentsToShow?.findIndex(
+    const index = postCommentsInfo?.findIndex(
       (c) => c.commentId === newCommentId,
     );
 
@@ -116,56 +147,53 @@ export default function PostScreen({
     });
 
     setNewCommentId(undefined);
-  }, [newCommentId, commentsToShow]);
+  }, [newCommentId, postCommentsInfo]);
 
-  useEffect(() => {
-    if (!isLoading && scrollToCommentId != null) {
-      setTimeout(() => {
-        const index = commentsToShow?.findIndex(
-          (c) => c.commentId === scrollToCommentId,
-        );
-
-        if (index == null || index < 0) {
-          return;
-        }
-
-        setTimeout(() => {
-          flashListRef.current?.scrollToIndex({
-            index,
-            animated: true,
-          });
-        }, 250);
-      }, 250);
-    }
-  }, [isLoading]);
+  const onRefresh = useCallback(() => {
+    refresh(async () => {
+      await instance.getComments({
+        post_id: postId,
+        community_id: postCommunityId ?? undefined,
+        sort: sortType,
+      });
+    });
+  }, []);
 
   if (postCommunityId == null) {
     return <LoadingScreen />;
   }
 
   return (
-    <YStack flex={1}>
-      <FlashList<ICommentInfo>
-        renderItem={renderItem}
-        data={commentsToShow}
-        keyExtractor={keyExtractor}
-        estimatedItemSize={100}
-        ListHeaderComponent={<Post />}
-        ListEmptyComponent={
-          <FeedLoadingIndicator
-            loading={
-              (isLoading && !isRefreshing) || awaitTransition.transitioning
-            }
-            error={isError}
-            empty={commentsToShow != null && commentsToShow.length < 1}
-          />
-        }
-        refreshControl={
-          <RefreshControl onRefresh={refresh} refreshing={isRefreshing} />
-        }
-        // @ts-expect-error this is valid
-        ref={flashListRef}
-      />
-    </YStack>
+    <PostScreenContext.Provider
+      value={{
+        refresh: onRefresh,
+        postCollapsed,
+        setPostCollapsed,
+      }}
+    >
+      <YStack flex={1}>
+        <FlashList<ICommentInfo>
+          renderItem={renderItem}
+          data={postCommentsInfo}
+          keyExtractor={keyExtractor}
+          estimatedItemSize={100}
+          ListHeaderComponent={<Post />}
+          ListEmptyComponent={
+            <FeedLoadingIndicator
+              loading={
+                (isLoading && !isRefreshing) || awaitTransition.transitioning
+              }
+              error={isError}
+              empty={postCommentsInfo != null && postCommentsInfo.length < 1}
+            />
+          }
+          refreshControl={
+            <RefreshControl onRefresh={onRefresh} refreshing={isRefreshing} />
+          }
+          // @ts-expect-error this is valid
+          ref={flashListRef}
+        />
+      </YStack>
+    </PostScreenContext.Provider>
   );
 }